tests/wpt/tests/html/semantics/scripting-1/the-script-element/script-type-whitespace.html

<!doctype html>
<title>&lt;script type> non-ASCII whitespace handling</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script>
function testParserInsertedDidNotRun(description) {
  test(() => assert_false(window.ran),
       "Script shouldn't run with " + description + " (parser-inserted)");
  window.ran = false;
}
</script>

<script>window.ran = false;</script>
<script type="text/javascript&#x000B;">window.ran = true;</script>
<script>testParserInsertedDidNotRun("type=\"text/javascript&#x000B;\"");</script>

<script type="text/javascript&#x0085;">window.ran = true;</script>
<script>testParserInsertedDidNotRun("type=\"text/javascript&#x0085;\"");</script>

<script type="text/javascript&#x00A0;">window.ran = true;</script>
<script>testParserInsertedDidNotRun("type=\"text/javascript&#x00A0;\"");</script>

<script type="text/javascript&#x1680;">window.ran = true;</script>
<script>testParserInsertedDidNotRun("type=\"text/javascript&#x1680;\"");</script>

<script type="text/javascript&#x3000;">window.ran = true;</script>
<script>testParserInsertedDidNotRun("type=\"text/javascript&#x3000;\"");</script>
